# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API",
        "description": "Description of your API",
        "contact": {
            "email": "ahmed@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/AddWallet": {
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Add a new wallet",
                "description": "Creates a new wallet for the authenticated user.",
                "operationId": "addWallet",
                "requestBody": {
                    "description": "Wallet data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "solde"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Savings"
                                    },
                                    "solde": {
                                        "type": "string",
                                        "example": "100.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wallet created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Wallet created successfully"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "wallet solde": {
                                            "type": "string"
                                        },
                                        "uuid": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erreur": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}